+++ core/cli_entry.py
@@ -5,11 +5,15 @@ from pprint import pformat
 import api
 import parsers
 import global_settings
+import progress
 
 from . import api_handler, parsing_handler
 
+EXIT_CODE_ERROR = 1
+EXIT_CODE_SUCCESS = 0
 
-def validate_and_upload_single_entry(directory):
+
+def validate_and_upload_single_entry(directory, force_upload=False):
     """
     This function acts as a single point of entry for uploading a directory
 
@@ -20,10 +24,31 @@ def validate_and_upload_single_entry(directory):
     Starts the upload
 
     :param directory: Directory of the sequencing run to upload
+    :param force_upload: When set to true, the upload status file will be ignored and file will attempt to be uploaded
     :return:
     """
     logging_start_block()
     logging.debug("validate_and_upload_single_entry:Starting {}".format(directory))
+
+    # Only upload if run is new, or force_upload is True
+    if not force_upload:
+        run_is_new = parsing_handler.run_is_new(directory)
+        if not run_is_new:
+            logging.error("ERROR! Run in directory {} is not new. It has either been uploaded, "
+                          "or an upload was attempted with error. "
+                          "Please check the status file 'irida_uploader_status.info' "
+                          "in the run directory for more details. "
+                          "You can bypass this error by uploading with the --force argument.".format(directory))
+            return exit_error()
+
+    # Add progress file to directory
+    try:
+        progress.write_directory_status(directory, progress.DIRECTORY_STATUS_PARTIAL)
+    except progress.exceptions.DirectoryError as e:
+        logging.error("ERROR! Error while trying to write status file to directory {} with error message: {}"
+                      "".format(e.directory, e.message))
+        logging.info("Samples not uploaded!")
+        return exit_error()
 
     # Do parsing (Also offline validation)
     try:
@@ -32,13 +57,13 @@ def validate_and_upload_single_entry(directory):
         # Directory was not valid for some reason
         logging.error("ERROR! An error occurred with directory '{}', with message: {}".format(e.directory, e.message))
         logging.info("Samples not uploaded!")
-        return exit_error()
+        return exit_error(directory)
     except parsers.exceptions.ValidationError as e:
         # Sequencing Run / SampleSheet was not valid for some reason
         logging.error("ERROR! Errors occurred during validation with message: {}".format(e.message))
         logging.error("Error list: " + pformat(e.validation_result.error_list))
         logging.info("Samples not uploaded!")
-        return exit_error()
+        return exit_error(directory)
 
     # Initialize the api for first use
     logging.info("*** Connecting to IRIDA ***")
@@ -48,7 +73,7 @@ def validate_and_upload_single_entry(directory):
         logging.error("ERROR! Could not initialize irida api.")
         logging.error("Errors: " + pformat(e.args))
         logging.info("Samples not uploaded!")
-        return exit_error()
+        return exit_error(directory)
     logging.info("*** Connected ***")
 
     logging.info("*** Verifying run (online validation) ***")
@@ -57,38 +82,58 @@ def validate_and_upload_single_entry(directory):
     except api.exceptions.IridaConnectionError as e:
         logging.error("Lost connection to Irida")
         logging.error("Errors: " + pformat(e.args))
-        return exit_error()
+        return exit_error(directory)
 
     if not validation_result.is_valid():
         logging.error("Sequencing run can not be uploaded")
         logging.error("Sequencing run can not be uploaded. Encountered {} errors"
                       "".format(validation_result.error_count()))
         logging.error("Errors: " + pformat(validation_result.error_list))
-        return exit_error()
+        return exit_error(directory)
     logging.info("*** Run Verified ***")
 
     # Start upload
     logging.info("*** Starting Upload ***")
     try:
-        api_handler.upload_sequencing_run(sequencing_run)
+        run_id = api_handler.upload_sequencing_run(sequencing_run)
     except api.exceptions.IridaConnectionError as e:
         logging.error("Lost connection to Irida")
         logging.error("Errors: " + pformat(e.args))
-        return 1
+        return exit_error(directory)
     logging.info("*** Upload Complete ***")
 
+    # Set progress file to complete
+    try:
+        progress.write_directory_status(directory, progress.DIRECTORY_STATUS_COMPLETE, run_id=run_id)
+    except progress.exceptions.DirectoryError as e:
+        # this is an exceptionally rare case (successful upload, but fails to write progress)
+        logging.ERROR("ERROR! Error while trying to write status file to directory {} with error message: {}"
+                      "".format(e.directory, e.message))
+        logging.info("Samples were uploaded, but progress file may be incorrect!")
+
     logging.info("Samples in directory '{}' have finished uploading!".format(directory))
+
+    logging_end_block()
 
     return exit_success()
 
 
-def exit_error():
+def exit_error(run_directory=None):
     """
     Returns an failed run exit code which ends the process when returned
+    :param run_directory: if this is given, the directory status will be set to error
     :return: 1
     """
+    if run_directory:
+        try:
+            progress.write_directory_status(run_directory, progress.DIRECTORY_STATUS_ERROR)
+        except progress.exceptions.DirectoryError as e:
+            # this is an exceptionally rare case (wrote to directory at start, but fails to write on exit)
+            logging.ERROR("ERROR! Error on exit while trying to write status file to directory {} "
+                          "with error message: {}".format(e.directory, e.message))
+
     logging_end_block()
-    return 1
+    return EXIT_CODE_ERROR
 
 
 def exit_success():
@@ -96,8 +141,7 @@ def exit_success():
     Returns an success run exit code which ends the process when returned
     :return: 0
     """
-    logging_end_block()
-    return 0
+    return EXIT_CODE_SUCCESS
 
 
 def logging_start_block():
@@ -109,6 +153,7 @@ def logging_start_block():
     logging.info("==================================================")
     logging.info("---------------STARTING UPLOAD RUN----------------")
     logging.info("Uploader Version {}".format(global_settings.UPLOADER_VERSION))
+    logging.info("Logging to file in: " + global_settings.log_file)
     logging.info("==================================================")