+++ core/parsing_handler.py
@@ -5,6 +5,7 @@ import logging
 
 import config
 import parsers
+from progress import upload_status
 
 from . import model_validator
 
@@ -50,3 +51,30 @@ def parse_and_validate(directory):
     logging.info("*** Parsing Done ***")
 
     return sequencing_run
+
+
+def run_is_new(directory):
+    """
+    Given a run directory, returns True or False if run is new
+    :param directory:
+    :return:
+    """
+    parser_instance = get_parser_from_config()
+    status = parser_instance.find_single_run(directory)
+    return _run_is_new_via_directory_status(status)
+
+
+def _run_is_new_via_directory_status(directory_status):
+    """
+    Checks if a run directory is new or not
+
+    :param directory_status: a DirectoryStatus object created by the progress module
+    :return: True or False
+    """
+    parser_instance = get_parser_from_config()
+    result = upload_status.get_directory_status(
+        directory_status.directory,
+        parser_instance.get_sample_sheet_file_name()
+    )
+
+    return result.status == upload_status.DIRECTORY_STATUS_NEW