+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changes
 =======
 
+Beta 0.2
+--------
+Added functionality:
+* When uploading, a `irida_uploader_status.info` status file will be created in the run directory
+* If a status file exists, trying to upload will fail and give a message saying an upload has already been attempted
+* using `--force` or `-f` when uploading bypasses this check
+
 Beta 0.1
 ---------
 * The start of time